move get_qualifications
